Invert home and IR pages
